Removed code for handling missing FrameworkInfo of a running task.
In previous versions of Mesos, the master might know about a task
running on an agent, but might not have a `FrameworkInfo` for that
task's framework. This might occur if the master fails over, the agent
re-registers, but the framework has not (yet) re-registered.

In Mesos 1.0, the agent will now supply the FrameworkInfo for all tasks
it is running when it re-registers with the master. Since we no longer
support 0.x agents, we can remove the old backward compatibility code
for handling a running task with unknown FrameworkInfo.

Note that this means that "orphan tasks", "orphan executors", and
"unregistered frameworks" are no longer possible.

Review: https://reviews.apache.org/r/59214