Validate DESTROY operation in `Resources::apply()`.
Ensure that applying `DESTROY` operation removes this volume. This is
needed to handle `DESTROY` of shared volumes, i.e. to  make sure we
correctly handle resources bookkeeping in the allocator so there is
only a single copy of the shared resource allocated if we allow the
operation to proceed.

Review: https://reviews.apache.org/r/59194/