Updated documentation for GPU support.
In order to enable the `gpu/nvidia` isolator, you must also enable the
`filesystem/linux` isolator. This has been enforced in the code for
quite some time (emitting an error if starting the agent without both
the proper combination set).

This commit now updates the documentation to explicity mention that
this is necessary and explain why.