Fixed flakiness in MasterAllocatorTest.ResourcesUnused.
In the test, `sched1` registers with the master and launches a task on
half the resources it is offered. Then `sched2` registers and should be
offered the remaining resources. The problem is that, because the test
did not pause the clock, a batch allocation might occur between the
first task launch and the second scheduler registering. The test tried
to account for this (by having `sched1` decline any additional offers it
receives), but this was not done correctly: declining the offer will
result in calling `recoverResources` again, which the test did not
account for.

Seems simpler to instead pause the clock, which should prevent the
possibility of a batch allocation in the first place.

Review: https://reviews.apache.org/r/59509