Added `os::copyfile(src, dst)` to stout.
The `os::copyfile()` method provides a cross-platform API to copy a file
from one location to another. The POSIX implementation shells out to
`cp`, and the Windows implementation uses the `::CopyFile` API. Note
that copying directories is not supported, and that the destination's
parent directory structure must exist (but the final path component must
be the destination file name, not a directory).

Review: https://reviews.apache.org/r/60621