Prevented allocating reservation refinements to non-capable frameworks.
When reservation refinements are present, old frameworks without the
RESERVATION_REFINEMENT capability won't be able to understand the new
format. While it's possible to translate the refined reservations into
the old format by "hiding" the intermediate reservations in the "stack",
this leads to ambiguity when processing RESERVE / UNRESERVE operations.
This is due to the loss of information when we drop the intermediate
reservations. Therefore, for now we simply filter out resources with
refined reservations if the framework does not have the capability.

Review: https://reviews.apache.org/r/60038