Added Config class to manage the config file in the new Mesos CLI.
This new class simplifies the management of the configuration file
given by the user; it loads the TOML file on initialization and
has one method for each element that the user can set.

This new class and its associated content is also passed to the
plugins at initialization so that they can read the user configuration
and use it.

Review: https://reviews.apache.org/r/59177/