Pulled http related declarations in the agent to a separate class.
Currently, it's nested in `Slave` class. This hurts readability a bit
and is not consistent with other structs like `Framework` or
`Executor`. This patch pulls the declarations into a separate Http
class and places it in a separate header to reduce the size of `Slave`
class.

Review: https://reviews.apache.org/r/59651