Optionally verify the source IP address for libprocess messages.
In general, libprocess does not validate that a peer is a
legitimate owner of the UPID it claims in a libprocess message.
This change adds a check that the IP address in the UPID matches
the peer address. This makes spoofing the UPID harder (e.g. to
send authenticated messages), but also breaks some legitimate
configurations, particularly on multihomed hosts.

Review: https://reviews.apache.org/r/58224/