Removed 'DO NOT USE' and 'EXPERIMENTAL' comments for multi-role support.
Now that the implementation of MULTI_ROLE is complete, we can
remove the comments that warn users not to use this functionality.

Also, since we've fixed some bugs at this point and there isn't any
missing functionality, this removes the EXPERIMENTAL comments as well.

Review: https://reviews.apache.org/r/59728