Added `path::from_uri` and `uri::from_path` to stout.
The former converts a well-formed URI to a path, fixing the path
separator for Windows.

The latter reverses the process, and produces a well-formed URI from a
path, keeping in mind that the URI path separator is always a forward
slash, and so converting from the Windows backslash path separator.

Review: https://reviews.apache.org/r/60622