Cleaned up use of 'this' pointers in dispatched callback.
By 'dispatch'ing to a PID the lines touched here already carefully
avoided calling member functions of a destructed object. This change
tidyies up the code to directly pass in the address of the member
function to call instead of constructing a lambda capturing around the
invocation which needs to capture 'this'.

Review: https://reviews.apache.org/r/59841/