CMake: Removed `ENABLE_OPTIMIZE` option.
This option was added as a mapping from the Autotools
`--enable-optimize` flag. However, it does not work as expected, since
CMake automatically adds platform-specific optimization flags depending
on the build configuration:
  * "Debug" builds are not optimized.
  * "Release" builds are optimized.

Therefore, all this option did was to unnecessarily add optimizations
to the Debug build, and fail to disable optimizations for the Release
build.  These configurations should be toggled:
  * For multi-configuration generators like MSVC:
    `cmake --build . --config [Debug|Release]`
  * For other generators:
    `cmake .. -DCMAKE_BUILD_TYPE=[Debug|Release]`

Review: https://reviews.apache.org/r/59772/