CMake: Removed `ENABLE_DEBUG` option.
This option was added as a mapping from the Autotools `--enable-debug`
flag. However, it does not work as expected, since some CMake generators
are multi-configuration. That is, CMake does not necessarily configure a
particular Debug or Release configuration, and instead (for some
generators, such as Visual Studio) it produces build files where the
congfiguration can be changed at build time.

The CMake documentation states that the variable `CMAKE_BUILD_TYPE`
should never be set in a CMake file (although it can be referenced, as
CMake is automatically invoked during builds, where it may be defined).
Thus multi-configuration generators did not respect what a user intended
with `-DENABLE_DEBUG=1` nor with `-DENABLE_DEBUG=0`, and the default
configuration was meaningless.

Review: https://reviews.apache.org/r/59773/