Updated Gauge to take std::function instead of Deferred.
A `Deferred` function object is always defined to be a
`std::function`, so when we are constructing a `Gauge` metrics
is it not strictly required for the callback to be a `Deferred`
object. This allows the user of Gauge to have `Process`-less
metrics, as well as avoiding the dispatch latency for `Process`
based metrics. However, the user must ensure that the function
call is safe up until the removal of the `Gauge` completes. The
pattern for how to do this for `Process`-based metrics has not
been established.

Review: https://reviews.apache.org/r/59858/