Adjusted (UN)RESERVE operations to handle reservation refinement [8/20].
The logic for applying RESERVE operations now needs to ensure that the
pre-reserved resources are being refined. In other words, the RESERVE
operation should be "pushing" a single reservation to the stack. Note
that we only support "pushing" a single reservation at a time.

For UNRESERVE, the operation should "pop" the last reservation off the
stack. We only support "popping" a single reservation at a time.

This patch also replaces all of the uses of `flatten()` to
`toUnreserved()` and also removes `flatten()` entirely. We also replace
the uses of `flatten(role, reservation)` except in tests. We'll remove
this version in later patch.

Review: https://reviews.apache.org/r/60021