Enabled authz on API calls which retrieve maintenance schedules.
Enables the use of authorization for the endpoint
'/maintenance/schedule' as well as the 'GET_MAINTENANCE_SCHEDULE' v1
API call, using the ACL 'GetMaintenanceSchedule' and the action of the
same name as the API call.

It also updates the ApiTests to take into account the authorization
case.

Review: https://reviews.apache.org/r/59100/