Enabled authorization when accesing maintenance status in v0 and v1 API.
Enables the use of authorization for the 'GET_MAINTENANCE_STATUS' v1
API call, as well as to the 'maintenance/status' endpoint, using the
ACL 'GetMaintenanceStatus' and the action of the same name as the API
call.

It also updates the ApiTests to take into account the authorization
case.

Review: https://reviews.apache.org/r/59101/