Enabled authorization for v1 calls starting and stopping maintenance.
Enables the use of authorization for the 'START_MAINTENANCE' and
'STOP_MAINTENANCE' v1 API calls, using the ACLs 'StartMaintenance' and
'StopMaintenance' respectively as well the actions of the same name as
the API calls.

It also updates the ApiTests to take into account the authorization
case.

Review: https://reviews.apache.org/r/59147/