Added protobuf changes for reservation refinement.
With reservation refinement, we now represent the reservation state in
`Resource.reservations` rather than with `Resource.role` and
`Resource.reservation`. We also keep track of the type of reservation,
and the role the reservation is for in the `Resource.ReservationInfo`.
This patch introduces these changes and explains the different formats
in detail.

Review: https://reviews.apache.org/r/59861/