Introduced `convertResourceFormat` to convert between resource formats.
With reservation refinement, we introduced a new resource format to
enable creating a refined reservation on top of an existing reservation.
We still have the "pre-reservation-refinement" format which uses
the `Resource.role` and `Resource.reservation` fields, and we now also
have the "post-reservation-refinement" format which uses
the `Resource.reservations` field to represent the reservation state.

In order to simplify our code, we use `convertResourceFormat` to
canonicalize the resources into the "post-reservation-refinement" in
memory, and convert them back as necessary.

Review: https://reviews.apache.org/r/60022