Updated the resources validation logic to allow for the endpoint format.
For backwards compatibility of external tools consuming endpoints,
We'll inject the "pre-reservation-refinement" format for resources
without refined reservations. That is, the `Resource.role` and
`Resource.reservation` fields will be present as well as
the in with the single `Resource.reservations` field.

Since the users of the endpoints are likely to feed the result back into
Mesos, we need to update the validation logic to allow these resources
to be considered valid.

Review: https://reviews.apache.org/r/60174