Avoided exposing `role` in JSON for resources with refined reservations.
With reservation refinement, we use the `Resource.reservations` field to
express the reservation state. Due to the fact that `Resource.role` is
an optional field __and__ has a default value, our generic Protobuf to
JSON code injects the `role` field with the default value even if it's
not set. Since this will likely confuse users, here we manually remove
the `Resource.role` field if the resource has refined reservations.

Review: https://reviews.apache.org/r/60036