Resources: Adjusted the predicates to the new resource format [2/20].
The predicate resource functions have the same preconditions as
the `Resources`, in that they require the given `Resource` object
to be validated, and in the "post-reservation-refinement" format.
As such, the corresponding preconditions have been added as a `CHECK`,
and the functions have been updated to inspect the correct fields.

Review: https://reviews.apache.org/r/60182