Introduced `downgradeResources` to facilitate out-bound resources.
This patch introduces `downgradeResources` utility which converts
resources in the "post-reservation-refinement" format to the
"pre-reservation-refinement" format. The function has all-or-nothing
semantics in that all are converted if none of the resources have
refined reservations and none are converted if any of the resources
have refined reservations.

Review: https://reviews.apache.org/r/60219