Fix bytes.hpp constexpr compilation failure with GCC 7.1.
Many of the `constexpr` variables fail to compile with errors such as
`error: 'Megabytes(32).Megabytes::<anonymous>' is not a constant
expression because it refers to an incompletely initialized variable`.
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=80829

This patch changes around the `Bytes` class a bit by getting rid of
the small classes that extend `Bytes`.

Review: https://reviews.apache.org/r/59413/