Placed the `convertResourceFormat` calls after resource validation.
`Resources` requires that the `Resource` objects being constructed with
be validated and are in the "post-reservation-refinement" format.
The `convertResourceFormat` was previously placed after the validation
of operations, but we construct `Resources` prior to operation
validation. This patch moves the conversion logic earlier to be done
right after resource validation.

Review: https://reviews.apache.org/r/60253