Prevented reserve/create with refined reservation on non-capable agents.
This patch makes it such that reserve / create operations involving
resources with refined reservations are invalid if they are attempted
on an agent without a RESERVATION_REFINEMENT capability. Such attempts
from an operator endpoint will result in a `BadRequest` response.
On the framework side, the operation will be dropped by the master.

Review: https://reviews.apache.org/r/60255