Avoided sending an inconsistent `TaskInfo` to the allocator.
Since we drop invalid tasks on a `LAUNCH` operation, we send a new
`Operation` with only the valid tasks appended. At the same time,
we adjust the `TaskInfo` a little bit to perform validation, and
to maintain backwards compatibility. However, the unmodified
`TaskInfo` was sent to the allocator. Since the allocator does not
look at the modified fields, this was inconsequential, but we should
avoid sending inconsistent `TaskInfo`s in general.

We also avoided creating an unnecessary copy of the `TaskInfo`.

Review: https://reviews.apache.org/r/60281