Introduced `validateAndUpgradeResources`.
Validation code for operations and tasks validated the resources as
the first step, and assumed valid resources from then on. This means
that it started to use functions such as `isDynamicallyReserved`.

However, since `isDynamicallyReserved` requires the resources to be
in the "post-reservation-refinement" format, we must convert the
resources before using those functions. For now, we introduce
a `validateAndUpgradeResources` abstraction which is called to validate
and convert the resources before invoking operation / task validation.

Review: https://reviews.apache.org/r/60282