Added Windows ReviewBot script.
This adds a support script used to run the Windows ReviewBot in a loop,
while properly logging the output and providing a URL for the log.

Unlike the other (Ubuntu) ReviewBot script, this script is standalone
and should be run on the machine that is performing the builds
(rather than given to Jenkins).

Review: https://reviews.apache.org/r/59500/