Fixed conversion warning in `duration.hpp`.
While `ns()` and our other types are `int64_t`, the variables to which
we're assigning are (by definition of `struct timeval`) a `time_t` or a
`long` (depending on platform). We cast it because there is no
alternative, and we otherwise generate hundreds of warnings. We cast to
`decltype(...)` because the type is different depending on platform.

Review: https://reviews.apache.org/r/59865/