Handle EINVAL from the capabilities version check.
capget will fail with the error EINVAL, and set the version field
to the kernel preferred value of _LINUX_CAPABILITY_VERSION_?
when an unsupported version value is specified.

So if we get EINVAL, we can treat that as success because we can
continue to verify the capabilities version.

Review: https://reviews.apache.org/r/60290/