Relaxed the resource format restriction for frameworks.
The current restriction is that the master will only accept resources
in the "post-reservation-refinement" format if the framework has opted
into the `RESERVATION_REFINEMENT` capability. This patch relaxes this
such that master will accept resources in any of the 3 formats.

For example, we output the "endpoint" format through the endpoints
for backward compatibility of external tooling. A framework should be
able to use the result of an endpoint and pass it back to Mesos,
since the result was produced by Mesos. This is especially applicable
to the V1 API. We also allow the "pre-reservation-refinement" format
because existing "resources files" are written in that format, and
they should still be usable without modification.

Review: https://reviews.apache.org/r/60352