Updated the `operator<<` for repeated resources to use `JSON::protobuf`.
The `Resource` objects in a `RepeatedPtrField<Resource>` may not be
validated and/or not converted to the "post-reservation-refinement"
format. Since `Resources` requires the resources to have been validated
and converted, we cannot rely on the conversion to be meaningful.
We opt to print the resources in their JSON representation instead.

Review: https://reviews.apache.org/r/60351