Add more linux/capabilities isolator test cases.
Add a case to verify that the effective framework capabilities
must be within the bounding framework capabilities.

Check that setting the framework capabilities to something that
is insufficient to execute ping overrides the operator flags and
fails.

Check that setting the framework bounding capabilities to allow
ping overrides the operator flags and succeeds.

Check that setting the framework effective and bounding
capabilities to allow ping overrides the operator flags and
succeeds.

Review: https://reviews.apache.org/r/60412/