Updated bundled ruby dependencies.
This commit updates to locked ruby dependencies in the 'Gemfile.lock'
to the lastest versions.

This e.g., updates eventmachine from 1.0.3 to 1.2.3. We need to
require a more recent version of eventmachine to be able to install
this bundle (which has native extension) on recent macOS.

Review: https://reviews.apache.org/r/60651/