Searched executables from the current working directory.
If a framework specifies use of its own executor (in the sandbox) and
sets shell to false, the executor is never found. This is becase if
the PATH env variable is set, `execvp` or `execvpe` won't look into
the current working directory for executables. This patch adds the
support for searching the current working directory for executables.

Review: https://reviews.apache.org/r/60280/