Fixed initialization of `__address__` in the abscense of `--ip` flag.
When we introduced the `__address6__` optional IPv6 storage into
`libprocess` we also introduced a regression because of which the `port`
doesn't get initialized to whatever is specified in the `--port` flag
until the `--ip` flag is specified.

This change fixes the initialization of the `__address__` in the
absence of the `--ip` flag.

Review: https://reviews.apache.org/r/60720/