Allowed to pass total resources in 'UpdateSlaveMessage'.
This commit both extends the existing 'UpdateSlaveMessage' proto
message, and adjusts its handling on the agent and master side.

To distinguish updates to 'oversubscribed_resources' from updates to
'total_resources' the message now contains a 'type' field to allow
disambiguation among an empty list of resources and an unset list of
resources. For backwards-compatibility we assume whenever the type
field was not set that caller intended to use the
'oversubscribed_resources' field as opposed to the 'total_resources'
field.

Currently, passing 'total_resources' is handled neither in the master
nor the default allocator; we will implement this in a subsequent
patch.

Review: https://reviews.apache.org/r/60641/