Fixed persistent volume and host volume conflict issue.
This is the fix for MESOS-7770. Basically, if a persistent volume
and a host volume are both specified and the source path of the
host volume is the same as the container path of the persistent
volume, then the persistent volume will be skipped and is not
mounted correctly. We should precisely check the mount table
to determine whether the persistent volume is mounted or not.
If not mounted, make sure we do mount the persistent volume.

Review: https://reviews.apache.org/r/60729/