Added double-checked locking for filter.
When running in"production" a filter will not be (likely) be set yet
with the current code the threads will experience head-of-line
blocking becaues they'll all queue up on the mutex. Test code will
still queue up but we don't care about the performance in these
situations!

Note that this patch also moves the filter into ProcessManager in the
effort towards eliminating globals.

Review: https://reviews.apache.org/r/60871