Added a method in the agent exposing a framework's allocated resources.
This allows us to call a method to get the framework's allocated
resources, instead of call-sites having to each iterate over the
framework's executors and the framework's pending tasks. It turns
out that most call-sites forgot to account for the pending tasks,
and accounting for the pending tasks is non-trivial!

Review: https://reviews.apache.org/r/60636/