Removed a misleading comment in the reservation endpoint tests.
This comment seemed to suggest that the reason for exposing
the reserved resources in the agent endpoint was that because
it is only *eventually* consistent with the master's state.
However, we expose agent state in both agent and master
endpoints because of the differing availability and performance
of the two components, as well as because the state is distributed
across the two (and we want to show the state at each distribution
point).