Updated agent webui page to display a reservation table.
This displays how many resources are reserved to which roles
(as well as displaying an entry for unreserved resources), as
well as how much of each reservation is allocated. It does not
show the roles that the resources are allocated to, which may
be different than the reservation role. In the case of
unreserved resources, the allocation role can be any role.

Review: https://reviews.apache.org/r/60370/