Introduced moodycamel/concurrentqueue as a 3rdparty library.
This is a header only library that provides a lock-free concurrent
queue which we can leverage in libprocess to improve performance.

The update to the CMake build will be added separately from this
change, note that this does not break the CMake build, since it
is not used yet. The plan is to have inclusion of concurrentqueue.h
be guarded by a configure flag to begin with, which means that it
will not break the CMake build, it will however prevent CMake
users from choosing to use the concurrent queue until the CMake
integration is complete.

Review: https://reviews.apache.org/r/60952