Factored out the event queue and made `state` be atomic.
This patch also introduces a function to construct a JSON
representation of a ProcessBase which is used to implement the
/__processes__ endpoint. We've then changed the implementation of that
endpoint to send messages to each process in order to have it inspect
it's event queue independently. In the future we may want to introduce
a high-priority queue to process events like these.

Review: https://reviews.apache.org/r/61055