CMake: Made `libprocess` use default linking.
The static/shared specifier defaults to `BUILD_SHARED_LIBS`,
which is defined as a CMake option (defaults to ON for Posix
and OFF for Windows).

The PIC property is redundant as it will be set via the global
`CMAKE_POSITION_INDEPENDENT_CODE` variable.  A separate CMake
file will define this appropriately when building shared libraries.

Review: https://reviews.apache.org/r/61284/