CMake: Removed unnecessary includes of `StoutConfigure`.
Because stout is now an "interface" library, the necessary metadata
is carried along with the stout target and does not need to be
spelled out by the downstream targets that use stout.  This means
that these downstream files do not need to include stout's CMake file.

Review: https://reviews.apache.org/r/61289/