Modified handling of parallel test configure flag for documentation.
This patch modifies the handling of the parallel test configure flag
for documentation purposes. With that we document proper ways to
handle both default-enabled and -disabled flags without using
arguments to 'AC_ARG_ENABLE'; the handling of '--disable-werror' is an
example of a default-enabled flag while
'--enable-parallel-test-execution' documents a default-disabled flag.

Review: https://reviews.apache.org/r/60646/